Revert to reactive property approach. Only provide decorators for packages that need coercion
